Ensure libthread stack aligned to 16 bytes for amd64
Ensure we'll still be aligned to 16 bytes after return address is pushed onto the stack.  Important because otherwise we'll get a GP exception if we try to use some SIMD instructions with a stack that's not aligned to 16 bytes.

This could be seen by building with gcc and running usb/usbd without any USB hubs.  An error case pushed the code into threadsysfatal, which had been compiled into some MOVAPS SIMD instructions, which expect the stack to be aligned to 16 bytes,

+++ amd64/include/u.h
@@ -25,7 +25,7 @@ typedef uint32_t	usize;
 typedef unsigned long size_t;
 typedef	uint32_t		Rune;
 typedef union FPdbleword FPdbleword;
-typedef uintptr		jmp_buf[10]; // for registers.
+typedef uint64_t	jmp_buf[10]; // for registers.
 
 #define	JMPBUFSP	6
 #define	JMPBUFPC	7
